ibloxapp/requirements.txt
psycopg2-binary
django==3.0
djangorestframework==3.1.1
djangorestframework_simplejwt